fix: resolve all make check failures
- Replace PEP 695 `type` statement in json_types.py with TypeAlias
  (pylint doesn't support the new syntax yet)
- Extract TestFileStore to separate test_file_store.py to fix
  too-many-lines (222/200) in test_demo_server.py
- Add demo optional dependency to dev group so fastapi/pydantic
  are available for tests
- Add missing AsyncMock for connect/disconnect in photo endpoint test
- Update vulture whitelist for new false positives
