+++ pyproject.toml
@@ -18,6 +18,7 @@ demo = [
 
 [dependency-groups]
 dev = [
+    "tg-auto-test[demo]",
     "pytest>=9.0.2,<10",
     "pytest-asyncio>=1.3.0,<2",
     "pytest-xdist>=3.8.0,<4",
